finish get and getById
