+++ controllers/contactsControllers.js
@@ -1,7 +1,5 @@
-const { Contact } = require("../db/contactModel");
-
 const {
-  // listContacts,
+  listContacts,
   getContactById,
   removeContact,
   addContact,
@@ -9,39 +7,22 @@ const {
   patchContact,
 } = require("../models/contacts");
 
-const { checkContactById } = require("../helpers/errors");
-
-// const getContactsController = async (req, res, next) => {
-//   const owner = req.user._id;
-//   const allContatcts = await Contact.find({ owner });
-//   console.log(allContatcts);
-//   res.json(allContatcts);
-// };
+const { checkContactById } = require("../helpers/checkById");
 
 const getContactsController = async (req, res) => {
   const { _id: owner } = req.user;
   const { page = 1, limit = 5, favorite = "" } = req.query;
   const skip = (page - 1) * limit;
-  if (favorite === "") {
-    const result = await Contact.find({ owner }, "-createdAt -updatedAt", {
-      skip,
-      limit,
-    }).populate("owner", "name email");
-    res.json(result);
-  } else {
-    const result = await Contact.find({ owner }, "-createdAt -updatedAt", {
-      skip,
-      limit,
-    })
-      .populate("owner", "name email")
-      .find({ favorite: { $eq: favorite } });
-    res.json(result);
-  }
+
+  const result = await listContacts(owner, favorite, skip, limit);
+
+  res.json(result);
 };
 
 const getContactByIdController = async (req, res, next) => {
   const { id } = req.params;
   const contactById = await getContactById(id);
+  console.log(contactById);
   await checkContactById(contactById);
   res.json({ contactById, message: "success" });
 };
@@ -52,11 +33,6 @@ const postContactController = async (req, res) => {
   const contact = await addContact({ ...req.body, owner });
   res.status(201).json({ message: "success", contact });
 };
-//   async (req, res, next) => {
-//   const { _id: owner } = req.user;
-//   const contact = await addContact(req.body, owner);
-//   res.status(201).json({ contact, message: "success" });
-// };
 
 const deleteContactController = async (req, res, next) => {
   const { id } = req.params;
